# [Misc][V1] Enhance performance of KVCacheManager._get_cached_block