Enhance dict retrieve by:
- fusing `in` check and `dict getitem` operation
- cache `dict.get` fn (cache the fn itself, rather than access `dict.get` every time) will get better performance
- use `for _ in dict.values()` for better performance than `next(iter(dict.values()))` because `for loop` use native `GETITER` and `NEXTITER` instruction and have better performance

## Benchmark

repeat 10000 times

| code | unknown block | cached block |
| - | - | - |
| main | 0.00107s | 0.00427s |
| [ver 0](#) | 0.00139s (+29.9%) | 0.00225s (-47.3%) |
| ver 1 (this PR) | 0.00114s (**+6.5%**) | 0.00179s (**-58.1%**) |